fix: separate RoomCategoryCreateRequest with RoomCategoriesCreateRequest
